Add plugin version compat w/ rules files
When loading rules, parse a required_plugin_versions item that looks
like:

- required_plugin_versions:
  - name: cloudtrail
    version: 1.0.1

And save it in the engine.

A new method is_plugin_compatible() takes a plugin and version and
returns whether it's compatible with all required_plugin_versions
blocks in all rules files.

In falco, after loading rules and plugins, iterate over the result of
plugin_infos(), calling is_plugin_compatible(), and exiting if any are
not compatible.